main/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.7
djangorestframework==3.14.0
idna==3.4
jwt==1.3.1
Pillow==10.1.0
pycparser==2.21
pycryptodome==3.19.0
python-weixin==0.5.0
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
xmltodict==0.13.0
